new cache script
