--- .github/workflows/anilist.yml
@@ -1,50 +1,0 @@
-name: anilist
-
-on:
-  workflow_dispatch:
-  # schedule:
-  #   # Once a week "At 00:00 on Sunday" (see https://crontab.guru/once-a-week)
-  #   - cron: "0 0 * * 0"
-
-jobs:
-  update-pool:
-    runs-on: ubuntu-latest
-    timeout-minutes: 180 # 3-hours
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          submodules: recursive
-          token: ${{ secrets.PAT }}
-          repository: ker0olos/fable
-
-      - uses: denoland/setup-deno@v1
-        with:
-          deno-version: v1.x
-      - run: deno run -A packs/anilist/pool.ts
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: anilist-pool
-          path: packs/anilist/pool.json
-
-      - uses: actions/upload-artifact@v4
-        with:
-          name: anilist-searching-directories
-          path: packs/anilist/*_directory.json
-
-      - name: Create Pull Request
-        uses: peter-evans/create-pull-request@v5
-        with:
-          token: ${{ secrets.PAT }}
-          author: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
-          title: '[bot] Updated "anilist/*.json"'
-          commit-message: "[bot] Updated `anilist/*.json`"
-          branch: update-anilist-pool
-          delete-branch: true
-          body: |
-            This PR is auto-generated
-
-            1. Updated the pool cache for each possible gacha query (See #9 #45)
-            2. Updated the searching directories for every character and media above `1000` in popularity (See #283)
-          labels: |
-            anilist
