+++ deno.json
@@ -15,7 +15,7 @@
         "no-external-import"
       ]
     },
-    "exclude": ["images-proxy", "dyn-images", "search-index"]
+    "exclude": ["images-proxy", "dyn-images"]
   },
   "lock": false,
   "tasks": {
@@ -31,7 +31,6 @@
     "exclude": [
       "images-proxy",
       "dyn-images",
-      "search-index",
       "packs/anilist/pool.json",
       "packs/anilist/characters_directory.json",
       "packs/anilist/media_directory.json"
@@ -48,6 +47,7 @@
     "sift": "https://raw.githubusercontent.com/ker0olos/sift/89e31b5a08af007d51073af31050fc2205245dd3/mod.ts",
     "tweetnacl": "https://esm.sh/tweetnacl@1.0.3",
     "mongodb": "npm:mongodb",
-    "mongodb-memory-server": "npm:mongodb-memory-server-core"
+    "mongodb-memory-server": "npm:mongodb-memory-server-core",
+    "orama": "npm:@orama/orama@2.0.11"
   }
 }
